requirements.txt
pandas==2.1.1
pyserial==3.5
